Rename/Refactor to prep for additional mod support.
Moved RemoteTechWrapper into a more generalized ModWrapper class to allow for re-use of reflection code.
